requirements.txt
fastmcp>=2.12.3
python-dotenv>=1.0
httpx>=0.28
